Renamed survey/institution/detail to survey/questiongroupmap/
