+++ apps/assessments/api_views/questions.py
@@ -7,6 +7,7 @@ from common.mixins import ILPStateMixin
 from common.models import Status
 from rest_framework.exceptions import ParseError
 
+from rest_framework.generics import ListAPIView
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework import viewsets
@@ -17,7 +18,8 @@ from django.db.models import Q
 from assessments.models import (
     QuestionGroup, Question, QuestionGroup_Questions,
     AnswerGroup_Institution, QuestionGroup_Institution_Association,
-    QuestionGroup_StudentGroup_Association, InstitutionImages
+    QuestionGroup_StudentGroup_Association, InstitutionImages,
+    Survey
 )
 from assessments.serializers import (
     QuestionGroupSerializer, QuestionSerializer,
@@ -100,7 +102,8 @@ class QGroupStoriesInfoView(ILPListAPIView):
                 questiongroup__survey__id=5).count(),
             'total_verified_stories': AnswerGroup_Institution.objects.filter(
                 questiongroup__survey__id=5).filter(is_verified=True).count(),
-            'total_images': InstitutionImages.objects.filter(answergroup__questiongroup__survey=5).count()
+            'total_images': InstitutionImages.objects.filter(
+                answergroup__questiongroup__survey=5).count()
         })
 
 
@@ -119,14 +122,18 @@ class QuestionGroupInstitutionAssociationViewSet(viewsets.ModelViewSet):
 
     def create(self, request, *args, **kwargs):
         if not self.request.data.get('questiongroup_ids'):
-            raise ParseError("Mandatory parameter questiongroup_ids not passed")
-        self.assessmentids = self.request.data.get('questiongroup_ids').split(",")
+            raise ParseError(
+                "Mandatory parameter questiongroup_ids not passed")
+        self.assessmentids = self.request.data.get(
+            'questiongroup_ids').split(",")
         if self.request.data.get('boundary_ids'):
             self.boundaryids = self.request.data.get('boundary_ids').split(",")
         if self.request.data.get('institution_ids'):
-            self.institutionids = self.request.data.get('institution_ids').split(",")
+            self.institutionids = self.request.data.get(
+                'institution_ids').split(",")
         if self.institutionids == [] and self.boundaryids == []:
-            raise ParseError("Mandatory parameter institution_ids or boundary_ids not passed")
+            raise ParseError(
+                "Mandatory parameter institution_ids/boundary_ids not passed")
         response = self.createAssessmentInstitutionAssociation()
 
         return response
@@ -184,3 +191,45 @@ class QuestionGroupStudentGroupAssociationViewSet(viewsets.ModelViewSet):
         headers = self.get_success_headers(serializer.data)
         return Response(
             serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+
+class SurveyQuestionGroupMapAPIView(ListAPIView, ILPStateMixin):
+
+    def list(self, request, *args, **kwargs):
+        survey_id = self.request.query_params.get('survey_id', None)
+        survey_on = Survey.objects.get(id=survey_id).survey_on.pk
+        institution_id = self.request.query_params.get('institution_id', None)
+        response = []
+        if survey_on == 'institution':
+            res = {}
+            qset = QuestionGroup_Institution_Association.objects.filter(
+                institution_id=institution_id,
+                questiongroup__survey_id=survey_id)
+            for qgroup_inst in qset:
+                res = {
+                    "id": qgroup_inst.questiongroup_id,
+                    "name": qgroup_inst.questiongroup.name,
+                    "type": qgroup_inst.questiongroup.type.char_id
+                }
+                response.append(res)
+        else:
+            res = {}
+            sg_qset = QuestionGroup_StudentGroup_Association.\
+                objects.filter(
+                    studentgroup__institution_id=institution_id,
+                )
+            for sgroup_inst in sg_qset:
+                sg_name = sgroup_inst.studentgroup.name
+                sg_id = sgroup_inst.studentgroup.id
+                res[sg_name] = {
+                    "id": sg_id, "name": sg_name
+                }
+                for studgroup_qgroup in sg_qset.filter(
+                        questiongroup__survey_id=survey_id):
+                    qgroup = studgroup_qgroup.questiongroup
+                    res[sg_name][qgroup.id] = {
+                        "id": qgroup.id, "name": qgroup.name
+                    }
+                    response.append(res)
+        return Response(response)
+
