+++ apps/assessments/api_views/survey.py
@@ -129,47 +129,6 @@ class SurveyInstitutionAPIView(ListAPIView, ILPStateMixin):
         return Response(response)
 
 
-class SurveyInstitutionDetailAPIView(ListAPIView, ILPStateMixin):
-
-    def list(self, request, *args, **kwargs):
-        survey_id = self.request.query_params.get('survey_id', None)
-        survey_on = Survey.objects.get(id=survey_id).survey_on.pk
-        institution_id = self.request.query_params.get('institution_id', None)
-        response = []
-        if survey_on == 'institution':
-            res = {}
-            qset = QuestionGroup_Institution_Association.objects.filter(
-                institution_id=institution_id,
-                questiongroup__survey_id=survey_id)
-            for qgroup_inst in qset:
-                res = {
-                    "id": qgroup_inst.questiongroup_id,
-                    "name": qgroup_inst.questiongroup.name,
-                    "type": qgroup_inst.questiongroup.type.char_id
-                }
-                response.append(res)
-        else:
-            res = {}
-            sg_qset = QuestionGroup_StudentGroup_Association.\
-                objects.filter(
-                    studentgroup__institution_id=institution_id,
-                )
-            for sgroup_inst in sg_qset:
-                sg_name = sgroup_inst.studentgroup.name
-                sg_id = sgroup_inst.studentgroup.id
-                res[sg_name] = {
-                    "id": sg_id, "name": sg_name
-                }
-                for studgroup_qgroup in sg_qset.filter(
-                        questiongroup__survey_id=survey_id):
-                    qgroup = studgroup_qgroup.questiongroup
-                    res[sg_name][qgroup.id] = {
-                        "id": qgroup.id, "name": qgroup.name
-                    }
-                    response.append(res)
-        return Response(response)
-
-
 class SurveyInstitutionAnsAggView(ListAPIView, ILPStateMixin):
     '''Returns all survey answers for a specific institution'''
     queryset = SurveyInstitutionQuestionGroupAnsAgg.objects.all()
