+++ apps/assessments/urls.py
@@ -13,7 +13,7 @@ from assessments.api_views import (
     AnswersStudentViewSet, SharedAssessmentsView, SurveyVolumeAPIView,
     SurveyClassQuestionKeyAPIView, SurveyQuestionGroupQuestionKeyAPIView,
     QuestionGroupSchoolViewSet, SurveyQuestionGroupDetailsAPIView,
-    SurveyInstitutionAnsAggView, SurveyInstitutionDetailAPIView,
+    SurveyInstitutionAnsAggView, SurveyQuestionGroupMapAPIView,
     SurveyTagAggAPIView, AssessmentsImagesView, AssessmentSyncView,
     RespondentTypeList, ShareYourStoryAPIView,
     QuestionGroupInstitutionAssociationViewSet,
@@ -180,9 +180,9 @@ urlpatterns = [
     url(r'surveys/storiesinfo/$',
         QGroupStoriesInfoView.as_view(),
         name='stories-info'),
-    url(r'survey/institution/detail/$',
-        SurveyInstitutionDetailAPIView.as_view(),
-        name='survey-institution-detial'),
+    url(r'survey/questiongroupmap/$',
+        SurveyQuestionGroupMapAPIView.as_view(),
+        name='survey-quesitongroup-map'),
     url(r'survey/summary/$',
         SurveySummaryAPIView.as_view(),
         name='survey-summary'),
